Recover the final particle update, camera setup, and field queries

+++ src/Event/evtMain.c
@@ -103,7 +103,7 @@ extern void *(*D_008873F4[])(size_t, size_t, u32);
 extern void (*jtbl_008873EC[])(void *ptr);
 void func_00286d80(u32 *, u32 *);
 s32 func_002853f0(s32 arg0);
-extern void func_00457140(s32 arg0, s32 arg1, s32 arg2, s32 arg3);
+extern void func_00457140(u8 arg0, u8 arg1, u8 arg2, u8 arg3);
 extern u8 D_0063C5C0[];
 void func_00442088();
 s32 func_0028f770(s32 arg0);
