+++ src/promoted/code1_0033.c
@@ -46,7 +46,7 @@ extern void func_003f6440(s32 arg0, s32 arg1);
 extern void func_002b8270(u8 *arg0, F2_0033 arg1, F2_0033 arg2,
                            u32 arg3, s32 arg4, s32 arg5);
 
-extern void func_00457140(s32 arg0, s32 arg1, s32 arg2, s32 arg3);
+extern void func_00457140(u8 arg0, u8 arg1, u8 arg2, u8 arg3);
 extern s8 iGpffffb59c;
 extern u8 *func_00452380(s8 *arg0);
 extern s32 func_0015f660(void);
